chore: moved all the sample_part_files fixtures into conftest.py
